feat: Add debug information to the response
Check for the debug request param and add the requested debug
information into the response body.
Refactor the ApeRender to make the code general purpose so it can be
used for both the meta data and the debug (and any other stuff that
might need to go in there in the future).